move qualspecies repo to repository.sbml.ext.qual
